Code quality and separation of concern
